+++ src/data/icons.ts
@@ -1,0 +1,6 @@
+// TODO: Add CUCAI links here
+export const icons = [
+    { src: "/email.png", link: "/" },
+    { src: "/instagram.png", link: "/" },
+    { src: "/linkedin.png", link: "/" },
+  ];
