+++ src/data/student_partners.ts
@@ -1,0 +1,70 @@
+export const student_partners = [
+    {
+      name: "WAT.ai",
+      logo: "/logos/wat-ai-logo.png",
+      university: "University of Waterloo",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 297,
+    },
+    {
+      name: "Western AI",
+      logo: "/logos/wai-logo.png",
+      university: "University of Western Ontario",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 120,
+    },
+    {
+      name: "McGill AI",
+      logo: "/logos/mcgill-ai-logo.png",
+      university: "McGill University",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 130,
+    },
+    {
+      name: "UofT AI",
+      logo: "/logos/uoftai-logo.png",
+      university: "University of Toronto",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 297,
+    },
+    {
+      name: "Wester Cyber Society",
+      logo: "/logos/wcs-logo.png",
+      university: "University of Western Ontario",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 297,
+      height: 100,
+    },
+    {
+      name: "UVic AI",
+      logo: "/logos/uvic-ai-logo.png",
+      university: "University of Victoria",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 150,
+      height: 100,
+    },
+    {
+      name: "UdeM AI",
+      logo: "/logos/udem-ai-logo.png",
+      university: "Universitié de Montréal",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 120,
+      height: 100,
+    },
+    {
+      name: "UTMIST",
+      logo: "/logos/utmist-logo.png",
+      university: "University of Toronto",
+      website: "https://watai.ca/",
+      insta: "https://www.instagram.com/wataiteam/",
+      width: 297,
+      height: 100,
+    },
+  ];
